mv .serverAssistedImport to Client
